create a borrower_frontend puppet module
This is the basic layout for a nested puppet module. The Puppetfile
inside the puppet/borrower_frontend folder is the definition of the
dependencies needed for the module.

This module can be pulled in to any using librarian-puppet by including
these lines in your Puppetfile:

```
mod 'LandRegistry/borrower_frontend',
  :git => 'https://github.com/LandRegistry/charges-borrower-frontend.git',
  :path => 'puppet/borrower_frontend'
```

and using it in your site.pp:

```
class profiles::developer_vm {
  class { 'borrower_frontend': }
}
```